modified:   Pipeline.py 	modified:   execute_rerun.py 	modified:   execute_run.py 	modified:   modules/Tribe_constructor.py
An error in the loading of the original catalog during reruns
would make the IDs on the tribe diferent to the events in reload of the
catalog. A new function in Trib_contructor.py allows load the Catalog
from the tribe, making sure to keep the IDs. It's currently being tested

Pipeline.py was modified to implement this function instead of reloading
 the catalog on reruns

execute_run.py and execute_rerun.py were modified to exclude the
partition gpu8xA100 since it's cpus are AMd and would cause some errors